Cosmetic improvements
