+++ Test_load.py
@@ -2,9 +2,8 @@
 
 import sys
 sys.path.append("B:\STUDIA MATERIALY\Gradient\Rozpoznawanie usmiechow\Rozpoznawanie usmiechow program")
-from server import load_data
+from data_server import *
 
-print("Hello World")
 filename = 'smile_warrior_dataset.csv'
 
 X_train, Y_train, X_test, Y_test, X_validate, Y_validate = load_data(filename)
@@ -12,6 +11,8 @@ X_train, Y_train, X_test, Y_test, X_validate, Y_validate = load_data(filename)
 print(X_train.shape, Y_train.shape)
 print(X_test.shape, Y_test.shape)
 print(X_validate.shape, Y_validate.shape)
+
+Show_Picture(3, X_train)
